default lat and lon as current position added
Attempt to fix the bug ; After adding the current location button,when changing the units (°F/°C) it was going back to the previous city displayed. I get that it happened because when changing from F° to °C, I call again the api based on the ${city} which is still set on the previous city used.  Now fixed thanks to useState of Lon and Lat